feat(window refres): window about to refresh
when refreshing the page, socket connection and twitter stream remains open.
data keep flowing in. that doesn't work with the flow of the app, so I added componentnWilUnmount, and added a listener for the window. before the compoenent unmount socket emit 'stop' and close the connection.

also changed the colors of the chart.